Added implicit classes to Truck, List, Point
Pimped List now allows list.swap or list.takeOff
Pimped Truck can now calculate the total cost, or swap stops between the trucks
Point list can now calculate the mean

Also refactored names to make it clearer what the traits for solution and truck
do.

The main problem still is the trucks have a copy of the location matrix.
This needs to change so only the solution has this. It can probably done neatly
with implicits.

I was also looking at not having stops in the trucks but  instead passing the links
in. This might not solve anything and put too much computation into the solution
classes.